+++ sample-web-base/src/main/java/com/sample/web/base/controller/html/BaseForm.java
@@ -40,5 +40,5 @@ public abstract class BaseForm implements Serializable {
      *
      * @return
      */
-    public abstract Integer getId();
+    public abstract Long getId();
 }
